fix: fixed model.eval() that was breaking hand_board batchnorm layer and causing hand misclassification. feat: added ability to compare shared weights
