+++ poker/models/model_layers.py
@@ -85,18 +85,6 @@ class NetworkFunctions(object):
 #              Processing Layers               #
 ################################################
 
-def compare_weights(net):
-    path = '/Users/morgan/Code/PokerAI/poker/checkpoints/frozen_layers/hand_board_weights'
-    if torch.cuda.is_available():
-        layer_weights = torch.load(path)
-    else:
-        layer_weights = torch.load(path,map_location=torch.device('cpu'))
-    print(net)
-    for name, param in net.named_parameters():
-        if name in layer_weights:
-            # print(param.data == layer_weights[name].data)
-            print(f'Layer {name},Equal {np.array_equal(param.data.numpy(),layer_weights[name].data.numpy())}')
-
 class ProcessHandBoard(nn.Module):
     def __init__(self,params,hand_length,hidden_dims=(16,32,32),activation_fc=F.relu):
         super().__init__()
@@ -130,9 +118,7 @@ class ProcessHandBoard(nn.Module):
         x: concatenated hand and board. alternating rank and suit.
         shape: B,M,18
         """
-        # compare_weights(self)
         B,M,C = x.size()
-        print('x',x)
         ranks,suits = unspool(x)
         # Shape of B,M,60,5
         hot_ranks = self.one_hot_ranks[ranks].to(self.device).float()
@@ -144,7 +130,6 @@ class ProcessHandBoard(nn.Module):
         for i in range(B):
             combinations = []
             for j in range(M):
-                # compare_weights(self)
                 s = self.suit_conv(hot_suits[i,j,:,:,:])
                 r = self.rank_conv(hot_ranks[i,j,:,:,:])
                 out = torch.cat((r,s),dim=-1)
@@ -154,14 +139,9 @@ class ProcessHandBoard(nn.Module):
                 out = self.categorical_output(out.view(60,-1))
                 # out = torch.argmax(torch.softmax(out,dim=-1))
                 combinations.append(torch.argmax(torch.softmax(out,dim=-1),dim=-1))
-                # print(f'Maximum value {torch.max(torch.softmax(out,dim=-1))}, Location {torch.argmax(torch.softmax(out,dim=-1))}')
             activations.append(torch.stack(combinations))
-        
         baseline = hardcode_handstrength(x)
-        # compare_weights(self)
         result = torch.stack(activations)
-        print('best hand guess',torch.min(result,dim=-1)[0])
-        print('baseline',baseline)
         return torch.min(result,dim=-1)[0].unsqueeze(-1)
         # return self.hand_out(torch.stack(activations).view(B,M,-1))
